prefix

+++ Bots/Raund1/Partners/Suppliers/DummySupplier.cs
@@ -11,6 +11,6 @@ namespace SpbAiChamp.Bots.Raund1.Partners.Suppliers
             IsFake = true;
         }
 
-        public override double CalculateCost(Consumer consumer) => Resource.HasValue ? 1 : int.MaxValue;
+        public override int CalculateCost(Consumer consumer) => Resource.HasValue ? 1 : int.MaxValue;
     }
 }
